Added merge / replace capability to save Template View.  added "press + to add new" and fixed when the last event is not cleared.  updated some syntax for modern objective c.  Added a timer to user defaults. removed need to persist viewDate. fixed possible problem with base time getting set weird by changedate

+++ NowWhat/Schedule.m
@@ -87,7 +87,12 @@
     
     NSError *error = nil;
     
-    NSNumber *count = [NSNumber numberWithInt:[moc countForFetchRequest:fetchRequest error:&error]];
+    NSNumber *count = [NSNumber numberWithInt:(int)([moc countForFetchRequest:fetchRequest error:&error])];
+    //NSUInteger *count = [moc countForFetchRequest:fetchRequest error:&error];
+    
+    
+    
+    
     
     NSLog(@"next schedule order is %@", count);
